move glue into glue/ dir
