+++ source/glue/glue.cpp
