+++ source/glue/glue.hpp
